+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/HostCompilerDirectives.java
@@ -166,26 +166,6 @@ public final class HostCompilerDirectives {
     }
 
     /**
-     * Marks a method that is called from a Truffle interpreter, but is not called frequently and is
-     * not important for interpreter performance.
-     * <p>
-     * This annotation is used to annotate methods that are called from a bytecode interpreter, but
-     * should generally not be inlined into the body of the bytecode interpreter. Language
-     * implementers are advised to inspect the IR of the interpreter when using this.
-     *
-     * @see BytecodeInterpreterSwitch to annotate the root method of a bytecode interpreter
-     *
-     * @deprecated use is no longer needed. boundaries for {@link BytecodeInterpreterSwitch} are
-     *             mostly determined automatically. To migrate remove all usages.
-     * @since 21.0
-     */
-    @Retention(RetentionPolicy.RUNTIME)
-    @Target({ElementType.METHOD, ElementType.CONSTRUCTOR})
-    @Deprecated(since = "22.2")
-    public @interface BytecodeInterpreterSwitchBoundary {
-    }
-
-    /**
      * Hints to Truffle host inlining that a particular method is partial evaluatable, but it would
      * be a good place for a cutoff when performing host inlining. A host compiler may use this
      * information as a hint to take trade-offs optimizing the code. Good examples of cutoffs are: